# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Fuel Nutrition API",
    "version": "1.0.0",
    "description": "Nutrition tracking API for the Fuel AI coach. Logs food, tracks weight, manages goals, and provides coaching insights."
  },
  "servers": [
    {
      "url": "https://fuel-2j8v.onrender.com/api",
      "description": "Production"
    }
  ],
...
  "paths": {
...
    "/food/analyze-photo": {
      "post": {
        "tags": ["Food"],
        "summary": "Analyze food photo with AI",
        "description": "Upload a food photo (base64 data URL) — Gemini Vision identifies food items and estimates macros. Returns identified foods with closest DB matches for logging.",
        "operationId": "analyzeFoodPhoto",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["image"],
                "properties": {
                  "image": {
                    "type": "string",
                    "description": "Base64 data URL of the food photo (e.g. data:image/jpeg;base64,...)."
                  },
                  "slot": {
                    "$ref": "#/components/schemas/MealSlot",
                    "description": "Optional meal slot to pre-select for logging."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Photo analysis result",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["items", "slot"],
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "required": ["name", "estimatedCalories", "estimatedProtein", "estimatedCarbs", "estimatedFat", "servingSize", "confidence", "matches"],
                        "properties": {
                          "name": { "type": "string" },
                          "estimatedCalories": { "type": "number" },
                          "estimatedProtein": { "type": "number" },
                          "estimatedCarbs": { "type": "number" },
                          "estimatedFat": { "type": "number" },
                          "servingSize": { "type": "string" },
                          "confidence": { "type": "number" },
                          "matches": {
                            "type": "array",
                            "items": {
                              "type": "object",
                              "properties": {
                                "id": { "type": "string" },
                                "name": { "type": "string" },
                                "calories": { "type": "number" },
                                "protein": { "type": "number" },
                                "carbs": { "type": "number" },
                                "fat": { "type": "number" },
                                "servingSize": { "type": "number" },
                                "servingUnit": { "type": "string" }
                              }
                            }
                          }
                        }
                      }
                    },
                    "slot": { "type": "string", "nullable": true }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid image format",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          },
          "500": {
            "description": "Analysis failed",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          }
        }
      }
    },
    "/food/search": {
      "get": {
        "tags": [
          "Food"
        ],
        "summary": "Search foods",
        "description": "Ranked food search over the seed DB + Indian foods + OpenFoodFacts + USDA + Spoonacular. Remote APIs fail soft. Returns up to `limit` matching foods.",
        "operationId": "searchFoods",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": ""
            },
            "description": "Free-text query. Empty returns seed DB foods only."
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 20,
              "minimum": 1,
              "maximum": 50
            },
            "description": "Max results."
          }
        ],
        "responses": {
          "200": {
            "description": "Search results",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FoodSearchResult"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/food/day": {
      "get": {
        "tags": [
          "Food"
        ],
        "summary": "Get food day view",
        "description": "Full day view with macro targets, consumed totals, remaining budgets, and meal groups with logged entries.",
        "operationId": "getFoodDay",
        "parameters": [
          {
            "name": "date",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date",
              "default": "2026-07-02"
            },
            "description": "ISO date (YYYY-MM-DD). Defaults to demo anchor date."
          }
        ],
        "responses": {
          "200": {
            "description": "Food day data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FoodDay"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Food"
        ],
        "summary": "Clear all food entries for a date",
        "description": "Removes every logged food entry for the given date. Returns count of cleared entries.",
        "operationId": "clearFoodDay",
        "parameters": [
          {
            "name": "date",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date",
              "default": "2026-07-02"
            },
            "description": "ISO date to clear."
          }
        ],
        "responses": {
          "200": {
            "description": "Day cleared",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "cleared",
                    "date",
                    "message"
                  ],
                  "properties": {
                    "cleared": {
                      "type": "integer",
                      "description": "Number of entries removed."
                    },
                    "date": {
                      "type": "string",
                      "format": "date"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/food/log": {
      "post": {
        "tags": [
          "Food"
        ],
        "summary": "Log a food",
        "description": "Log a food item against a day and meal slot. Returns the updated FoodDay for that date.",
        "operationId": "logFood",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LogFoodRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Food logged — returns updated day",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FoodDay"
                }
              }
            }
          },
          "400": {
            "description": "Invalid body or foodId not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/food/log/{id}": {
      "delete": {
        "tags": [
          "Food"
        ],
        "summary": "Delete a logged food entry",
        "description": "Remove a logged food entry by its log ID. Returns the updated FoodDay for the date.",
        "operationId": "deleteLoggedFood",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "LoggedFood ID (e.g. 'log-1')."
          },
          {
            "name": "date",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date",
              "default": "2026-07-02"
            },
            "description": "ISO date of the day the entry belongs to."
          }
        ],
        "responses": {
          "200": {
            "description": "Entry deleted — returns updated day",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FoodDay"
                }
              }
            }
          },
          "404": {
            "description": "Entry ID not found for that date",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/food/copy": {
      "post": {
        "tags": [
          "Food"
        ],
        "summary": "Copy food entries between dates",
        "description": "Copy all food entries from one date to another ('copy yesterday' flow). Returns the updated FoodDay for the destination date.",
        "operationId": "copyFoodDay",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "from",
                  "to"
                ],
                "properties": {
                  "from": {
                    "type": "string",
                    "format": "date",
                    "description": "Source date (YYYY-MM-DD)."
                  },
                  "to": {
                    "type": "string",
                    "format": "date",
                    "description": "Destination date (YYYY-MM-DD)."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Copied successfully — returns updated destination day",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FoodDay"
                }
              }
            }
          },
          "400": {
            "description": "Missing from or to",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/food/barcode": {
      "get": {
        "tags": ["Food"],
        "summary": "Lookup food by barcode",
        "description": "Returns a single food item from OpenFoodFacts by barcode (EAN/UPC). Use to log packaged foods by scanning barcode.",
        "operationId": "getFoodByBarcode",
        "parameters": [
          {
            "name": "code",
            "in": "query",
            "required": true,
            "schema": { "type": "string" },
            "description": "Barcode (EAN-13, UPC-A, etc.)"
          }
        ],
        "responses": {
          "200": {
            "description": "Food item found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "item": { "$ref": "#/components/schemas/FoodItem" }
                  }
                }
              }
            }
          },
          "404": { "description": "Barcode not found" }
        }
      }
    },
    "/food/recent": {
      "get": {
        "tags": ["Food"],
        "summary": "Get frequently logged foods",
        "description": "Returns the top-N most frequently logged foods. Optionally filtered by meal slot (slot bonus: +3 weight) and recency (maxDays).",
        "operationId": "getRecentFoods",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": { "type": "integer", "default": 10 }
          },
          {
            "name": "slot",
            "in": "query",
            "required": false,
            "schema": { "$ref": "#/components/schemas/MealSlot" },
            "description": "Boost foods logged in this meal slot."
          },
          {
            "name": "maxDays",
            "in": "query",
            "required": false,
            "schema": { "type": "integer" },
            "description": "Only count entries from the last N days."
          }
        ],
        "responses": {
          "200": {
            "description": "Recent foods list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "foodId": { "type": "string" },
                          "name": { "type": "string" },
                          "count": { "type": "integer" },
                          "lastDate": { "type": "string", "format": "date" },
                          "calories": { "type": "number" },
                          "protein": { "type": "number" },
                          "carbs": { "type": "number" },
                          "fat": { "type": "number" }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
...
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string",
            "description": "Error message."
          }
        }
      },
...
      "MealSlot": {
        "type": "string",
        "enum": [
          "breakfast",
          "lunch",
          "dinner",
          "snack"
        ],
        "description": "Meal slot for grouping food entries."
      },
      "FoodSource": {
        "type": "string",
        "enum": [
          "database",
          "custom",
          "recipe"
        ],
        "description": "Where a food definition originates."
      },
      "Macros": {
        "type": "object",
        "required": [
          "calories",
          "protein",
          "carbs",
          "fat"
        ],
        "properties": {
          "calories": {
            "type": "number",
            "description": "Energy in kcal."
          },
          "protein": {
            "type": "number",
            "description": "Protein in grams."
          },
          "carbs": {
            "type": "number",
            "description": "Carbohydrates in grams."
          },
          "fat": {
            "type": "number",
            "description": "Fat in grams."
          }
        }
      },
      "FoodItem": {
        "type": "object",
        "required": [
          "id",
          "name",
          "source",
          "servingSize",
          "servingUnit",
          "calories",
          "protein",
          "carbs",
          "fat"
        ],
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique food identifier (e.g. 'off-123', 'db-chicken-breast')."
          },
          "name": {
            "type": "string",
            "description": "Food display name."
          },
          "brand": {
            "type": "string",
            "description": "Optional brand or restaurant name."
          },
          "source": {
            "$ref": "#/components/schemas/FoodSource"
          },
          "servingSize": {
            "type": "number",
            "description": "Numeric size of one serving (e.g. 100 for 100g)."
          },
          "servingUnit": {
            "type": "string",
            "description": "Unit label (e.g. 'g', 'cup', 'scoop', 'slice')."
          },
          "calories": {
            "type": "number",
            "description": "Kcal per serving."
          },
          "protein": {
            "type": "number",
            "description": "Protein in g per serving."
          },
          "carbs": {
            "type": "number",
            "description": "Carbs in g per serving."
          },
          "fat": {
            "type": "number",
            "description": "Fat in g per serving."
          },
          "barcode": {
            "type": "string",
            "description": "Optional barcode for scan-to-log."
          },
          "favorite": {
            "type": "boolean",
            "description": "Whether user has starred this item."
          }
        }
      },
      "FoodSearchResult": {
        "type": "object",
        "required": [
          "query",
          "items"
        ],
        "properties": {
          "query": {
            "type": "string"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FoodItem"
            }
          }
        }
      },
      "LoggedFood": {
        "allOf": [
          {
            "$ref": "#/components/schemas/Macros"
          },
          {
            "type": "object",
            "required": [
              "id",
              "date",
              "slot",
              "foodId",
              "name",
              "quantity",
              "servingUnit",
              "loggedAt"
            ],
            "properties": {
              "id": {
                "type": "string",
                "description": "Unique log entry ID (e.g. 'log-1')."
              },
              "date": {
                "type": "string",
                "format": "date",
                "description": "Date logged."
              },
              "slot": {
                "$ref": "#/components/schemas/MealSlot"
              },
              "foodId": {
                "type": "string",
                "description": "Reference to the FoodItem."
              },
              "name": {
                "type": "string"
              },
              "brand": {
                "type": "string"
              },
              "quantity": {
                "type": "number",
                "description": "Number of servings."
              },
              "servingUnit": {
                "type": "string"
              },
              "loggedAt": {
                "type": "string",
                "format": "date-time",
                "description": "ISO timestamp of when it was logged."
              }
            }
          }
        ]
      },
      "MealGroup": {
        "type": "object",
        "required": [
          "slot",
          "label",
          "entries",
          "totals"
        ],
        "properties": {
          "slot": {
            "$ref": "#/components/schemas/MealSlot"
          },
          "label": {
            "type": "string",
            "description": "Human-readable label (e.g. 'Breakfast', 'Lunch')."
          },
          "entries": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LoggedFood"
            }
          },
          "totals": {
            "$ref": "#/components/schemas/Macros",
            "description": "Rolled-up macros for this meal slot."
          }
        }
      },
      "FoodDay": {
        "type": "object",
        "required": [
          "date",
          "target",
          "consumed",
          "remaining",
          "groups"
        ],
        "properties": {
          "date": {
            "type": "string",
            "format": "date"
          },
          "target": {
            "$ref": "#/components/schemas/Macros",
            "description": "Daily macro targets."
          },
          "consumed": {
            "$ref": "#/components/schemas/Macros",
            "description": "Sum of all logged entries."
          },
          "remaining": {
            "$ref": "#/components/schemas/Macros",
            "description": "Target minus consumed."
          },
          "groups": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MealGroup"
            },
            "description": "Four meal slot groups (breakfast, lunch, dinner, snack)."
          }
        }
      },
      "LogFoodRequest": {
        "type": "object",
        "required": [
          "date",
          "slot",
          "foodId",
          "quantity"
        ],
        "properties": {
          "date": {
            "type": "string",
            "format": "date",
            "description": "ISO date to log against."
          },
          "slot": {
            "$ref": "#/components/schemas/MealSlot"
          },
          "foodId": {
            "type": "string",
            "description": "Must exist in the food DB or remote cache."
          },
          "quantity": {
            "type": "number",
            "exclusiveMinimum": 0,
            "description": "Number of servings."
          }
        }
      },
...
    }
  }
}